% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttcalib-package.R
\docType{package}
\name{ttcalib-package}
\alias{ttcalib-package}
\alias{_PACKAGE}
\title{ttcalib: Calibration of travel times to empirical data}
\description{
Calibration of travel times to empirical data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UrbanAnalyst/ttcalib}
  \item Report bugs at \url{https://github.com/UrbanAnalyst/ttcalib/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
