% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{ttcalib_centrality}
\alias{ttcalib_centrality}
\title{Load an 'SC' street network, weight for motorcar routing, and calculate
centralityl.}
\usage{
ttcalib_centrality(path)
}
\arguments{
\item{path}{Path to 'SC'-format file containing street network data.}
}
\value{
Network with centrality estimates on each edge.
}
\description{
Load an 'SC' street network, weight for motorcar routing, and calculate
centralityl.
}
