% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodata.R
\name{ttcalib_geodata}
\alias{ttcalib_geodata}
\title{Read geometries for chosen citie from Uber Movement data.}
\usage{
ttcalib_geodata(path)
}
\arguments{
\item{path}{Path to directory containing Uber movement data.}
}
\value{
A 'data.frame' with centroids of polygons used to aggregate movement
data.
}
\description{
Currently hard-coded to Brussels data only.
}
