% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate-penalties.R
\name{ttcalib_penalties}
\alias{ttcalib_penalties}
\title{Calculate standard error against 'uber' data for all weighted networks
generated via \link{ttcalib_streetnet_batch}.}
\usage{
ttcalib_penalties(path)
}
\arguments{
\item{path}{Path to 'SC'-format file containing street network data.}
}
\description{
Calculate standard error against 'uber' data for all weighted networks
generated via \link{ttcalib_streetnet_batch}.
}
