% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveltimes.R
\name{ttcalib_traveltimes}
\alias{ttcalib_traveltimes}
\title{Estimate travel times between all Uber movement polygons}
\usage{
ttcalib_traveltimes(graph, geodata, uberdata)
}
\arguments{
\item{graph}{A \pkg{dodgr} graph returned from \link{ttlib_centrality}.}

\item{geodata}{A \code{data.frame} returned from \link{ttlib_geodata}.}

\item{uberdata}{A \code{data.frame} returned from \link{ttlib_uberdata}.}
}
\value{
A 'data.frame' with columns of \pkg{m4ra} estimates of travel times
and corresponding empirical values from Uber movement data.
}
\description{
Estimate travel times between all Uber movement polygons
}
