% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uberdata.R
\name{ttcalib_uberdata}
\alias{ttcalib_uberdata}
\title{Read Uber Movement data and filter to defined time limits.}
\usage{
ttcalib_uberdata(path, hours = NULL)
}
\arguments{
\item{path}{Path to directory containing Uber movement data.}

\item{hours}{A vector of two values defining the range of hours for data to
be filtered. Default of \code{NULL} returns aggregate of all hours without
filtering.}
}
\value{
A 'data.frame' of Uber Movement estimates of travel times.
}
\description{
Currently hard-coded to Brussels data only.
}
