% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate-centrality.R
\name{ttcalib_centrality}
\alias{ttcalib_centrality}
\title{Calibrate travel times to network centrality.}
\usage{
ttcalib_centrality(
  path_graph,
  path_uberdata,
  city = "santiago",
  hours = c(7, 10),
  turn_penalty = 1
)
}
\arguments{
\item{path_graph}{Path to locally-saved \code{fst}-format weighted street
network including centrality column.}

\item{path_uberdata}{Path to Uber movement data.}

\item{city}{Currently only accepts "santiago"}

\item{hours}{A vector of two values defining the range of hours for Uber
Movement data to be filtered. Value of \code{NULL} aggregates all hours without
filtering.}

\item{turn_penalty}{The value of the time penalty for waiting to turn across
oncoming traffic used to generate the graph stored at \code{path}.}
}
\description{
This calibration step is performed after calibration to waiting-time
penalties for traffic lights and turning across oncoming traffic, performed
with the \link{ttcalib_penalties} function. For Santiago, that function gives
an optimal waiting time at traffic lights of 16 seconds, and waiting time to
turn across oncoming traffic of 1 second. These values should be used to
generate a weighted network with time-based centrality via the
\link{ttcalib_streetnet} function.
}
\details{
The result of that call is presumed to have been saved using the \code{fst}
package (with \code{write_fst}), which strips all attributes of the graph. These
attributes must then be manually re-instated, and so are required to be
submitted as parameters to
this function.
}
