% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate-penalties.R
\name{ttcalib_penalties}
\alias{ttcalib_penalties}
\title{Calculate standard error against 'uber' data for all weighted networks
generated via \link{ttcalib_streetnet_batch}.}
\usage{
ttcalib_penalties(path, city, hours = NULL)
}
\arguments{
\item{path}{Path to directory containing Uber movement data.}

\item{city}{One of "brussels" or "santiago" (case-insensitive).}

\item{hours}{A vector of two values defining the range of hours for data to
be filtered. Default of \code{NULL} returns aggregate of all hours without
filtering.}
}
\description{
Calculate standard error against 'uber' data for all weighted networks
generated via \link{ttcalib_streetnet_batch}.
}
