% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetnet.R
\name{ttcalib_streetnet}
\alias{ttcalib_streetnet}
\title{Load an 'SC' street network, weight for motorcar routing, and optionally
calculate centrality.}
\usage{
ttcalib_streetnet(
  path,
  centrality = FALSE,
  penalty_traffic_lights = 8,
  penalty_turn = 7.5
)
}
\arguments{
\item{path}{Path to 'SC'-format file containing street network data.}

\item{centrality}{If \code{TRUE}, calculate network centrality on all graph edges.
Load an 'SC' street network, weight for motorcar routing, and calculate
centrality.}

\item{penalty_traffic_lights}{Time penalty for waiting at traffic lights (in
seconds).}

\item{penalty_turn}{Time penalty for turning across oncoming traffic.}
}
\value{
Network with centrality estimates on each edge.
}
\description{
Load an 'SC' street network, weight for motorcar routing, and optionally
calculate centrality.
}
